pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastauth"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "raushanraja", email = "vraushanraja@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
 "fastapi[standard]==0.112.2,<0.114.0",
 "PyJWT==2.9.0",
 "passlib[bcrypt]==1.7.4",
 "python-dotenv==1.0.1",
 "uvicorn==0.30.6",
 "httpx==0.27.2",
 "pydantic-settings==2.4.0"
]

[project.urls]
Documentation = "https://github.com/raushanraja/fastauth#readme"
Issues = "https://github.com/raushanraja/fastauth/issues"
Source = "https://github.com/raushanraja/fastauth"

[tool.hatch.version]
path = "src/fastauth/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/fastauth tests}"

[tool.hatch.envs.default.scripts]
app = "python src/fastauth/main.py"


[tool.coverage.run]
source_pkgs = ["fastauth", "tests"]
branch = true
parallel = true
omit = [
  "src/fastauth/__about__.py",
]

[tool.coverage.paths]
fastauth = ["src/fastauth", "*/fastauth/src/fastauth"]
tests = ["tests", "*/fastauth/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
